requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==2.0.29
alembic==1.13.1
python-dotenv==1.0.1
asyncpg==0.29.0
fastapi_pagination
loguru==0.7.2
